+++ rasa/shared/core/training_data/story_reader/story_reader.py
@@ -58,10 +58,26 @@ class StoryReader:
 
     @staticmethod
     def is_test_stories_file(filename: Text) -> bool:
+        """Checks if the specified file is a test story file.
+
+        Args:
+            filename: File to check.
+
+        Returns:
+            `True` if specified file is a test story file, `False` otherwise.
+        """
         raise NotImplementedError
 
     @staticmethod
     def is_stories_file(filename: Text) -> bool:
+        """Checks if the specified file is a story file.
+
+        Args:
+            filename: File to check.
+
+        Returns:
+            `True` if specified file is a story file, `False` otherwise.
+        """
         raise NotImplementedError
 
     def _add_current_stories_to_result(self):
